Add Parser to extract message for Encoder.
Parser class with a minimal API to parse a message
and return the contents of either a string or a
file passed in via a single array argument.

Parser#ensure_correct_number_of_arguments checks
for array argument length and returns an error
symbol if array contains either more or less than
the single expected arg.

If no error, extract message to be encoded from 
the argument array from the file or string arg.

Include tests to cover string, file, and empty
arguments in parsing. Include tests for the
parsing outcome.