Add Obfuscator class to convert any encoded morse code into an obfuscated sequence of characters.
Obfuscator#obfuscate converts an input message 
into an obfuscated message by first replacing all
dots, and then by replacing all dashes with a 
simple algorithm-based series of characters.

Include tests for Obfuscator class. Test for the
obfuscation of a valid character sequence as well
as an invalid sequence.

Test Obfuscator#obfuscate_with which is used for
both dot and dash obfuscation. Include a test for
invalid characters for both dots and dashes.